Add ability to log into AO3 automatically
This means that you can generate feeds for works that are only available
to registered users.

Please note that you should not expose this service logged in publically.
I run this on my phone, restricted to localhost.

This is currently restricted to the rocket implementation, but that
won't last long as I need another way of getting the username/password
*anyway*.